Fazendo modificações - Modificando Atributo "Promocao"
